+++ packages/shared/index.ts
@@ -21,7 +21,7 @@ export * from './syncRef'
 export * from './throttledWatch'
 export * from './toReactive'
 export * from './toRefs'
-export * from './tryOnBeforeUnmount'
+export * from './tryOnBeforeUnmount_'
 export * from './tryOnMounted'
 export * from './tryOnScopeDispose'
 export * from './tryOnUnmounted'
